+++ node-red-contrib-ignition/CHANGELOG.md
@@ -5,3 +5,29 @@
 September 25, 2017
 
  * Initial Release
+ * Compatible with Ignition 7.9
+ 
+### 1.5.6
+
+July 20, 2020
+
+ * Updated for Ignition 8.0
+ * Compatible with Ignition 8.0
+ 
+### 1.5.7
+
+July 30, 2020
+
+ * Allows a write of 0 or false in the tag write node
+ * Compatible with Ignition 8.1
+ * Compatible with Ignition Maker
+ 
+### 1.5.8
+
+August 15, 2021
+
+ * Added web socket tag read node
+ * Added browse tag node
+ * Added support for reading/writing multiple tags
+ * Compatible with Ignition 8.1
+ * Compatible with Ignition Maker